Refactor the data migration part in a separate module
